refactor(wardrobe): split ClothingFormScreen into focused files
916-line god file broken into three focused files:
- FormComponents.kt — DropdownSelector<T>, ColorSelectionGrid, DatePickerField,
  BrandAutocompleteField, isColorDark
- FormPreviews.kt — all @Preview functions + single source of truth for
  preview mock data (internal vals)
- ClothingFormScreen.kt — screen entry point, ClothingFormTopBar,
  ClothingFormContent only (401 lines, down from 916)

Adds architecture-roadmap.md tracking remaining refactor work.
